+++ packages/article-skeleton/src/article-body/article-body.web.js
@@ -23,7 +23,10 @@ import {
   InArticlePuff,
   InlineNewsletterPuff,
   PreviewNewsletterPuff,
-  AutoNewsletterPuff
+  AutoNewsletterPuff,
+  OptaFootballFixtures,
+  OptaFootballStandings,
+  OptaFootballSummary
 } from "@times-components/ts-components";
 
 import ArticleLink from "./article-link";
@@ -169,17 +172,41 @@ const renderers = ({
           />
         );
 
+      case "opta-football-fixtures-v3":
+        return (
+          <OptaFootballFixtures
+            season={attributes.season}
+            competition={attributes.competition}
+            date_from={attributes["date-from"]}
+            date_to={attributes["date-to"]}
+          />
+        );
+
+      case "opta-football-standings-v3":
+        return (
+          <OptaFootballStandings
+            season={attributes.season}
+            competition={attributes.competition}
+            navigation
+          />
+        );
+
+      case "opta-football-match-summary-v3":
+        return (
+          <OptaFootballSummary
+            season={attributes.season}
+            competition={attributes.competition}
+            match={attributes.match}
+          />
+        );
+
       case "in-article-puff":
         if (inArticlePuffFlag) {
           return (
             <Context.Consumer key={key}>
               {({ theme }) => (
                 <FetchProvider url={deckApiUrl + attributes["deck-id"]}>
-                  <InArticlePuff
-                    sectionColour={theme.sectionColour}
-                    forceImageAspectRatio="3:2"
-                    sanitiseHtml
-                  />
+                  <InArticlePuff sectionColour={theme.sectionColour} />
                 </FetchProvider>
               )}
             </Context.Consumer>