+++ packages/ts-components/src/components/in-article-link/styles.ts
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+import { colours, fonts } from '@times-components/styleguide';
+
+export const Link = styled.a`
+  display: flex;
+  flex-direction: row;
+  margin-top: 16px;
+  color: #bf0000;
+
+  fill: #bf0000;
+
+  :hover {
+    color: ${colours.functional.secondary};
+    fill: ${colours.functional.secondary};
+  }
+`;
+
+export const LinkText = styled.div`
+  margin-top: 1px;
+  margin-right: 12px;
+  font-family: ${fonts.supporting};
+  font-size: 16px;
+`;