+++ packages/article-main-standard/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.33.8](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.33.7...@times-components/article-main-standard@3.33.8) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.33.7](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.33.6...@times-components/article-main-standard@3.33.7) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.33.6](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.33.5...@times-components/article-main-standard@3.33.6) (2021-06-04)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.33.5](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.33.4...@times-components/article-main-standard@3.33.5) (2021-06-04)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
 ## [3.33.4](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.33.3...@times-components/article-main-standard@3.33.4) (2021-05-27)
 
 **Note:** Version bump only for package @times-components/article-main-standard